requirements.txt
langchain
langchain-ollama
langchain-milvus
langchain-chroma
pandas
pymilvus
fastapi
uvicorn 
python-dotenv
pydantic-settings
tiktoken
borneo
aiofiles
pyhive
thrift
google-cloud-translate==3.*
openai
httpx 
PyJWT
python-multipart
google-cloud-aiplatform
google-genai
langid
motor==3.6.0
pytest==8.3.2
pytest-asyncio==0.23.8
asgi-lifespan==2.1.0
pydantic[email]
playwright
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
requests>=2.31.0
umap-learn>=0.5.5
